Refactoring the folder and file structure
