+++ src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import App from './App';
+import App from './components/App';
 import registerServiceWorker from './registerServiceWorker';
 import './styles/index.css';
 
@@ -10,4 +10,7 @@ ReactDOM.render(
   <App  />,
   document.getElementById('root') as HTMLElement
 );
+
+
+
 registerServiceWorker();
